Non-blocking code work in process.
Working on the ability to flag a menu item as being selected so that the
menu system will flow into that item's callback function.  The callback
functions then will be written as non-blocking which will allow more
dynamic systems to be built.

Functions are not working at this time, but I wanted to migrate the code
to another machine to keep working on it.